refactor: 쿠링봇 인터페이스/구현체 이름을 추상화
