+++ data/remote/src/main/java/com/ku_stacks/ku_ring/remote/kuringbot/KuringBotClient.kt
@@ -1,6 +1,6 @@
-package com.ku_stacks.ku_ring.remote.sse
+package com.ku_stacks.ku_ring.remote.kuringbot
 
-interface SSEDataSource {
+interface KuringBotClient {
     /**
      * 쿠링봇 SSE 세션을 열고, 데이터를 문자열 형식으로 받아 처리한다.
      * SSE 데이터를 받았을 때, [onReceived]에 접두어 `data:`를 떼지 않고 그대로 제공한다.
